`testIntegration`: Rename `programs` to `tests`
