+++ src/testIntegration/kotlin/tests/TypeExpressionTests.kt
@@ -1,0 +1,11 @@
+package tests
+
+import kotlin.test.Test
+
+class TypeExpressionTests {
+    @Test
+    fun testUnresolved() {
+        // Test that a type expression which references non-existing entities
+        // results in a semantic error
+    }
+}
