+++ src/testIntegration/resources/tests/euler/problem10.sigma
