+++ src/testIntegration/resources/tests/euler/problem6.sigma
