+++ src/testIntegration/resources/tests/euler/problem7.sigma
