+++ src/testIntegration/resources/tests/euler/problem8.sigma
