+++ src/testIntegration/resources/tests/euler/problem9.sigma
